requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
djangorestframework==3.17.1
idna==3.11
joblib==1.5.3
numpy==2.4.3
pandas==3.0.1
python-dateutil==2.9.0.post0
requests==2.33.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
